use "uintptr_t*" for buffers in VASurfaceAttributeTPI
it supports both 32bits and 64bits system, either pointer or handle/fd

+++ va/va_tpi.h
@@ -24,6 +24,7 @@
 #ifndef _VA_TPI_H_
 #define _VA_TPI_H_
 
+#include <stdint.h>
 #include <va/va.h>
 
 #ifdef __cplusplus
@@ -60,7 +61,7 @@ typedef struct _VASurfaceAttributeTPI {
     unsigned int chroma_u_offset; /* U offset from the beginning of the memory */
     unsigned int chroma_v_offset; /* V offset from the beginning of the memory */
     unsigned int count; /* buffer count for surface creation */
-    unsigned int *buffers; /* buffer handles or user pointers */
+    uintptr_t *buffers; /* buffer handles or user pointers */
     unsigned int reserved[4]; /* used to pass additional information, like
                                * Android native window pointer
                                */
